refactor: rename internal components for clarity and migrate variableDeclarationName helper
