+++ pkg/compiler/internal/expr_query.go
@@ -1,0 +1,301 @@
+package internal
+
+import (
+	"strings"
+
+	"github.com/MontFerret/ferret/v2/pkg/bytecode"
+	"github.com/MontFerret/ferret/v2/pkg/compiler/internal/core"
+	"github.com/MontFerret/ferret/v2/pkg/parser/diagnostics"
+	"github.com/MontFerret/ferret/v2/pkg/parser/fql"
+	"github.com/MontFerret/ferret/v2/pkg/runtime"
+	"github.com/MontFerret/ferret/v2/pkg/source"
+)
+
+func (c *ExprCompiler) compileQueryExpression(ctx fql.IQueryExpressionContext) bytecode.Operand {
+	if ctx == nil {
+		return bytecode.NoopOperand
+	}
+
+	plan := collectRecoveryPlan(c.ctx, ctx, core.RecoveryPlanOptions{})
+	return c.ctx.PolicyCompiler.CompileWithRecoveryPlan(plan, core.CatchJumpModeNone, func() bytecode.Operand {
+		if ctx == nil {
+			return bytecode.NoopOperand
+		}
+
+		src, ok := c.compileQueryExpressionSource(ctx)
+		if !ok {
+			return bytecode.NoopOperand
+		}
+
+		span := diagnostics.SpanFromRuleContext(ctx)
+		modifier := queryModifierName(ctx.QueryModifier())
+		queryReg := c.emitQueryEnvelope(ctx, span)
+		queryResult := c.emitApplyQuery(span, src, queryReg)
+		dst := c.lowerQueryModifier(span, modifier, queryResult)
+
+		if dst.IsRegister() {
+			c.ctx.Types.Set(dst, queryResultTypeForModifier(modifier))
+		}
+
+		return dst
+	})
+}
+
+func (c *ExprCompiler) compileQueryExpressionSource(ctx fql.IQueryExpressionContext) (bytecode.Operand, bool) {
+	if ctx == nil {
+		return bytecode.NoopOperand, false
+	}
+
+	sourceExpr := ctx.Expression()
+	if sourceExpr == nil {
+		return bytecode.NoopOperand, false
+	}
+
+	source := c.Compile(sourceExpr)
+	return source, source != bytecode.NoopOperand
+}
+
+func (c *ExprCompiler) emitQueryEnvelope(ctx fql.IQueryExpressionContext, span source.Span) bytecode.Operand {
+	queryReg := c.ctx.Registers.Allocate()
+
+	c.ctx.Emitter.WithSpan(span, func() {
+		c.ctx.Emitter.EmitArray(queryReg, 3)
+	})
+
+	kind := c.compileQueryKindOperand(ctx)
+	c.emitQueryEnvelopeOperand(span, queryReg, kind)
+
+	payload := c.compileQueryPayloadOperand(ctx.QueryPayload())
+	c.emitQueryEnvelopeOperand(span, queryReg, payload)
+
+	options := c.compileQueryOptionsOperand(ctx.QueryWithOpt())
+	c.emitQueryEnvelopeOperand(span, queryReg, options)
+
+	return queryReg
+}
+
+func (c *ExprCompiler) compileQueryKindOperand(ctx fql.IQueryExpressionContext) bytecode.Operand {
+	kind := ""
+	if ident := ctx.GetDialect(); ident != nil {
+		kind = strings.ToLower(ident.GetText())
+	}
+
+	return loadConstant(c.ctx, runtime.NewString(kind))
+}
+
+func (c *ExprCompiler) compileQueryPayloadOperand(ctx fql.IQueryPayloadContext) bytecode.Operand {
+	if ctx == nil {
+		return loadConstant(c.ctx, runtime.EmptyString)
+	}
+
+	if literal := ctx.StringLiteral(); literal != nil {
+		if value, ok := parseStringLiteralConst(literal); ok {
+			return loadConstant(c.ctx, value)
+		}
+
+		return c.ctx.LiteralCompiler.CompileStringLiteral(literal)
+	}
+
+	if param := ctx.Param(); param != nil {
+		return c.CompileParam(param)
+	}
+
+	if variable := ctx.Variable(); variable != nil {
+		return c.CompileVariable(variable)
+	}
+
+	return loadConstant(c.ctx, runtime.EmptyString)
+}
+
+func (c *ExprCompiler) compileQueryOptionsOperand(ctx fql.IQueryWithOptContext) bytecode.Operand {
+	if ctx == nil || ctx.Expression() == nil {
+		return loadConstant(c.ctx, runtime.None)
+	}
+
+	return c.Compile(ctx.Expression())
+}
+
+func (c *ExprCompiler) emitQueryEnvelopeOperand(span source.Span, queryReg, value bytecode.Operand) {
+	c.ctx.Emitter.WithSpan(span, func() {
+		c.ctx.Emitter.EmitArrayPush(queryReg, value)
+	})
+}
+
+func (c *ExprCompiler) emitApplyQuery(span source.Span, src, queryReg bytecode.Operand) bytecode.Operand {
+	result := c.ctx.Registers.Allocate()
+
+	c.ctx.Emitter.WithSpan(span, func() {
+		c.ctx.Emitter.EmitABC(bytecode.OpQuery, result, src, queryReg)
+	})
+
+	return result
+}
+
+func (c *ExprCompiler) lowerQueryModifier(span source.Span, modifier queryModifier, queryResult bytecode.Operand) bytecode.Operand {
+	switch modifier {
+	case queryModifierExists:
+		dst := c.ctx.Registers.Allocate()
+		c.ctx.Emitter.WithSpan(span, func() {
+			c.ctx.Emitter.EmitAB(bytecode.OpExists, dst, queryResult)
+		})
+		return dst
+	case queryModifierCount:
+		dst := c.ctx.Registers.Allocate()
+		c.ctx.Emitter.WithSpan(span, func() {
+			c.ctx.Emitter.EmitAB(bytecode.OpLength, dst, queryResult)
+		})
+		return dst
+	case queryModifierAny:
+		dst := c.ctx.Registers.Allocate()
+		zero := c.ctx.Symbols.AddConstant(runtime.NewInt(0))
+		c.ctx.Emitter.WithSpan(span, func() {
+			c.ctx.Emitter.EmitABC(bytecode.OpLoadIndexOptionalConst, dst, queryResult, zero)
+		})
+		return dst
+	case queryModifierValue:
+		return c.ctx.lowerQueryModifierValue(span, queryResult)
+	case queryModifierOne:
+		return c.ctx.lowerQueryModifierOne(span, queryResult)
+	default:
+		return queryResult
+	}
+}
+
+func queryModifierName(ctx fql.IQueryModifierContext) queryModifier {
+	if ctx == nil {
+		return queryModifierUnknown
+	}
+
+	return parseQueryModifier(ctx.GetText())
+}
+
+func queryResultTypeForModifier(modifier queryModifier) core.ValueType {
+	switch modifier {
+	case queryModifierExists:
+		return core.TypeBool
+	case queryModifierCount:
+		return core.TypeInt
+	case queryModifierAny, queryModifierValue, queryModifierOne:
+		return core.TypeAny
+	default:
+		return core.TypeList
+	}
+}
+
+func parseQueryModifier(text string) queryModifier {
+	switch strings.ToLower(text) {
+	case string(queryModifierExists):
+		return queryModifierExists
+	case string(queryModifierCount):
+		return queryModifierCount
+	case string(queryModifierAny):
+		return queryModifierAny
+	case string(queryModifierValue):
+		return queryModifierValue
+	case string(queryModifierOne):
+		return queryModifierOne
+	default:
+		return queryModifierUnknown
+	}
+}
+
+func (c *CompilerContext) lowerQueryModifierValue(span source.Span, queryResult bytecode.Operand) bytecode.Operand {
+	dst := c.Registers.Allocate()
+	cond := c.Registers.Allocate()
+	zero := c.Symbols.AddConstant(runtime.NewInt(0))
+	message := c.Symbols.AddConstant(runtime.NewString(queryValueFailMessage))
+	success := c.Emitter.NewLabel("query", string(queryModifierValue), "ok")
+	end := c.Emitter.NewLabel("query", string(queryModifierValue), "end")
+
+	c.Emitter.WithSpan(span, func() {
+		c.Emitter.EmitAB(bytecode.OpExists, cond, queryResult)
+		c.Emitter.EmitJumpIfTrue(cond, success)
+		c.Emitter.EmitLoadNone(dst)
+		c.Emitter.EmitA(bytecode.OpFail, message)
+		c.Emitter.EmitJump(end)
+		c.Emitter.MarkLabel(success)
+		c.Emitter.EmitABC(bytecode.OpLoadIndexConst, dst, queryResult, zero)
+		c.Emitter.MarkLabel(end)
+	})
+
+	return dst
+}
+
+func (c *CompilerContext) lowerQueryModifierOne(span source.Span, queryResult bytecode.Operand) bytecode.Operand {
+	dst := c.Registers.Allocate()
+	length := c.Registers.Allocate()
+	one := c.Symbols.AddConstant(runtime.NewInt(1))
+	zero := c.Symbols.AddConstant(runtime.NewInt(0))
+	message := c.Symbols.AddConstant(runtime.NewString(queryOneFailMessage))
+	success := c.Emitter.NewLabel("query", string(queryModifierOne), "ok")
+	end := c.Emitter.NewLabel("query", string(queryModifierOne), "end")
+
+	c.Emitter.WithSpan(span, func() {
+		c.Emitter.EmitAB(bytecode.OpLength, length, queryResult)
+		c.Emitter.EmitJumpCompare(bytecode.OpJumpIfEqConst, length, one, success)
+		c.Emitter.EmitLoadNone(dst)
+		c.Emitter.EmitA(bytecode.OpFail, message)
+		c.Emitter.EmitJump(end)
+		c.Emitter.MarkLabel(success)
+		c.Emitter.EmitABC(bytecode.OpLoadIndexConst, dst, queryResult, zero)
+		c.Emitter.MarkLabel(end)
+	})
+
+	return dst
+}
+
+func (c *ExprCompiler) compileQueryLiteral(ctx fql.IQueryLiteralContext) bytecode.Operand {
+	if ctx == nil {
+		return bytecode.NoopOperand
+	}
+
+	kind := ""
+	if ident := ctx.Identifier(); ident != nil {
+		kind = strings.ToLower(ident.GetText())
+	}
+
+	dst := c.ctx.Registers.Allocate()
+	span := diagnostics.SpanFromRuleContext(ctx)
+
+	c.ctx.Emitter.WithSpan(span, func() {
+		c.ctx.Emitter.EmitArray(dst, 3)
+	})
+
+	kindReg := loadConstant(c.ctx, runtime.NewString(kind))
+
+	c.ctx.Emitter.WithSpan(span, func() {
+		c.ctx.Emitter.EmitArrayPush(dst, kindReg)
+	})
+
+	payloadReg := loadConstant(c.ctx, runtime.EmptyString)
+	if str := ctx.StringLiteral(); str != nil {
+		if val, ok := parseStringLiteralConst(str); ok {
+			payloadReg = loadConstant(c.ctx, val)
+		} else {
+			payloadReg = c.ctx.LiteralCompiler.CompileStringLiteral(str)
+		}
+	}
+
+	c.ctx.Emitter.WithSpan(span, func() {
+		c.ctx.Emitter.EmitArrayPush(dst, payloadReg)
+	})
+
+	params := ctx.Expression()
+	var paramsReg bytecode.Operand
+
+	if params == nil {
+		paramsReg = loadConstant(c.ctx, runtime.None)
+	} else {
+		paramsReg = c.Compile(params)
+	}
+
+	c.ctx.Emitter.WithSpan(span, func() {
+		c.ctx.Emitter.EmitArrayPush(dst, paramsReg)
+	})
+
+	if dst.IsRegister() {
+		c.ctx.Types.Set(dst, core.TypeAny)
+	}
+
+	return dst
+}
